Add Item controller, Item model, basic file structure (moved files into folders) and added mysqlconnection.py
